pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Denis Tretyakov <den.tretyakoff@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
taskipy = "^1.13.0"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
start = "python manage.py runserver 0:8000"
